Split JsonConfig into shim and non-shim files.
